refactor package based on feature
