+++ app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".feature.login.view.LoginActivity"></activity>
+        <activity android:name="com.nikhil.mybestfriend.feature.auth.login.view.LoginActivity"></activity>
         <activity android:name=".feature.splash.view.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
