--- app/src/main/java/com/nikhil/mybestfriend/feature/api/interceptor/ConnectivityInterceptor.kt
@@ -1,6 +1,0 @@
-package com.nikhil.mybestfriend.feature.api.interceptor
-
-import okhttp3.Interceptor
-
-interface ConnectivityInterceptor : Interceptor {
-}
