--- app/src/main/java/com/nikhil/mybestfriend/feature/api/interceptor/ConnectivityInterceptorImpl.kt
@@ -1,24 +1,0 @@
-package com.nikhil.mybestfriend.feature.api.interceptor
-
-import android.content.Context
-import android.net.ConnectivityManager
-import com.nikhil.mybestfriend.feature.exception.NoConnectionException
-import okhttp3.Interceptor
-import okhttp3.Response
-
-class ConnectivityInterceptorImpl(context: Context) : ConnectivityInterceptor {
-    private val appContext = context.applicationContext
-
-    override fun intercept(chain: Interceptor.Chain): Response {
-        if (!isNetworkAvailable())
-            throw NoConnectionException()
-        return chain.proceed(chain.request())
-    }
-
-    private fun isNetworkAvailable(): Boolean {
-        val connectivityManager = appContext.getSystemService(Context.CONNECTIVITY_SERVICE)
-                as ConnectivityManager
-        val networkInfo = connectivityManager.activeNetworkInfo
-        return networkInfo != null && networkInfo.isConnected
-    }
-}
