+++ app/src/main/java/com/nikhil/mybestfriend/feature/auth/login/view/LoginActivity.kt
@@ -1,0 +1,13 @@
+package com.nikhil.mybestfriend.feature.auth.login.view
+
+import android.os.Bundle
+import com.nikhil.mybestfriend.R
+import com.nikhil.mybestfriend.feature.commons.view.BaseActivity
+
+class LoginActivity : BaseActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_login)
+    }
+}
