--- app/src/main/java/com/nikhil/mybestfriend/feature/base/view/BaseActivity.kt
@@ -1,8 +1,0 @@
-package com.nikhil.mybestfriend.feature.base.view
-
-import androidx.appcompat.app.AppCompatActivity
-
-open class BaseActivity : AppCompatActivity() {
-
-
-}
