--- app/src/main/java/com/nikhil/mybestfriend/feature/base/viewmodel/BaseViewModel.kt
@@ -1,4 +1,0 @@
-package com.nikhil.mybestfriend.feature.base.viewmodel
-
-open class BaseViewModel {
-}
