+++ app/src/main/java/com/nikhil/mybestfriend/feature/cat/data/api/response/CatWeight.kt
@@ -1,0 +1,7 @@
+package com.nikhil.mybestfriend.feature.cat.data.api.response
+
+data class CatWeight(
+    val imperial: String,
+    val metric: String
+)
+
