+++ app/src/main/java/com/nikhil/mybestfriend/feature/cat/data/db/entity/CatDetails.kt
@@ -1,0 +1,4 @@
+package com.nikhil.mybestfriend.feature.cat.data.db.entity
+
+class CatDetails {
+}
