--- app/src/main/java/com/nikhil/mybestfriend/feature/cat/model/CatDetails.kt
@@ -1,26 +1,0 @@
-package com.nikhil.mybestfriend.feature.cat.model
-
-import com.google.gson.annotations.SerializedName
-
-
-data class CatDetails(
-    val approved: Int,
-    @SerializedName("breed_ids")
-    val breedIds: String,
-    val breeds: List<CatBreed>,
-    @SerializedName("created_at")
-    val createdAt: String,
-    val height: Int,
-    val id: String,
-    @SerializedName("original_filename")
-    val originalFilename: String,
-    val pending: Int,
-    val rejected: Int,
-    @SerializedName("sub_id")
-    val subId: String,
-    val url: String,
-    val width: Int
-)
-
-
-
