--- app/src/main/java/com/nikhil/mybestfriend/feature/cat/model/CatWeight.kt
@@ -1,7 +1,0 @@
-package com.nikhil.mybestfriend.feature.cat.model
-
-data class CatWeight(
-    val imperial: String,
-    val metric: String
-)
-
