--- app/src/main/java/com/nikhil/mybestfriend/feature/cat/repo/CatDetailDataSource.kt
@@ -1,14 +1,0 @@
-package com.nikhil.mybestfriend.feature.cat.repo
-
-import androidx.lifecycle.LiveData
-import com.nikhil.mybestfriend.feature.cat.model.CatDetails
-
-interface CatDetailDataSource {
-
-    val catDetailList: LiveData<List<CatDetails>>
-    suspend fun fetchCatBreed(
-        breedIds: String,
-        includeBreeds: Boolean = true
-    )
-
-}
