+++ app/src/main/java/com/nikhil/mybestfriend/feature/commons/data/api/interceptor/ConnectivityInterceptor.kt
@@ -1,0 +1,6 @@
+package com.nikhil.mybestfriend.feature.commons.data.api.interceptor
+
+import okhttp3.Interceptor
+
+interface ConnectivityInterceptor : Interceptor {
+}
