+++ app/src/main/java/com/nikhil/mybestfriend/feature/commons/exception/NoConnectionException.kt
@@ -1,0 +1,8 @@
+package com.nikhil.mybestfriend.feature.commons.exception
+
+import java.lang.RuntimeException
+
+class NoConnectionException : RuntimeException() {
+    override val message: String?
+        get() = "Unable to connect to internet"
+}
