+++ app/src/main/java/com/nikhil/mybestfriend/feature/commons/view/BaseActivity.kt
@@ -1,0 +1,8 @@
+package com.nikhil.mybestfriend.feature.commons.view
+
+import androidx.appcompat.app.AppCompatActivity
+
+open class BaseActivity : AppCompatActivity() {
+
+
+}
