+++ app/src/main/java/com/nikhil/mybestfriend/feature/commons/view/BaseFragment.kt
@@ -1,0 +1,18 @@
+package com.nikhil.mybestfriend.feature.commons.view
+
+import android.os.Bundle
+import android.view.View
+import androidx.fragment.app.Fragment
+
+open  class BaseFragment : Fragment() {
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        initFragment()
+    }
+
+    open protected fun initFragment()  {
+
+    }
+
+
+}
