+++ app/src/main/java/com/nikhil/mybestfriend/feature/commons/viewmodel/BaseViewModel.kt
@@ -1,0 +1,6 @@
+package com.nikhil.mybestfriend.feature.commons.viewmodel
+
+import androidx.lifecycle.ViewModel
+
+open class BaseViewModel : ViewModel() {
+}
