--- app/src/main/java/com/nikhil/mybestfriend/feature/exception/NoConnectionException.kt
@@ -1,8 +1,0 @@
-package com.nikhil.mybestfriend.feature.exception
-
-import java.lang.RuntimeException
-
-class NoConnectionException : RuntimeException() {
-    override val message: String?
-        get() = "Unable to connect to internet"
-}
