+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.login.view.LoginActivity">
+    tools:context=".feature.auth.login.view.LoginActivity">
 
     <fragment
         android:id="@+id/fragment"
