+++ app/src/main/res/layout/fragment_cat_detail.xml
@@ -4,7 +4,7 @@
    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".feature.cat.view.fragment.CatDetailFragment">
+    tools:context=".feature.cat.details.view.CatDetailFragment">
 
 
    <ImageView
