+++ app/src/main/res/layout/fragment_cat_list.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/dogSwipeRefreshLayout"
-    tools:context=".feature.cat.view.fragment.CatListFragment"
+    tools:context=".feature.cat.listing.view.CatListFragment"
     >
 
     <androidx.constraintlayout.widget.ConstraintLayout
